+++ package/sdk/src/core.ts
@@ -10,4 +10,4 @@ import { ClientSDK } from "./lib/sdks.js";
  * application and passed around through some dependency injection mechanism  to
  * parts of an application that need to make SDK calls.
  */
-export class PetstoreCore extends ClientSDK {}
+export class PetsCore extends ClientSDK {}